+++ src/Controller/ArticulosController.php
@@ -2,7 +2,7 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
-
+use Cake\ORM\TableRegistry;
 /**
  * Articulos Controller
  *
@@ -24,6 +24,10 @@ class ArticulosController extends AppController
      */
     public function index()
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+            return $this->redirect($this->referer());
+        }
         $this->paginate = [
             'contain' => ['Modelos']
         ];
@@ -52,6 +56,29 @@ class ArticulosController extends AppController
      */
     public function view($id = null)
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            $asig=TableRegistry::get('Asignaciones')->find('all')
+                ->where(['articulo_id ='=>$id])
+                ->andWhere(['hasta >='=>date('Y-m-d')]);
+            if($asig->isEmpty()){
+                $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+                return $this->redirect($this->referer());
+            }
+            $found=false;
+            foreach ($asig as $row){
+                $pro_tra=TableRegistry::get('ProcesosTrabajadores')->find('all')
+                        ->where(['proceso_id ='=>$row->proceso_id])
+                        ->andWhere(['trabajador_id ='=>$this->request->session()->read('Auth.User.trabajador_id')]);
+                if($pro_tra!=null&&!$pro_tra->isEmpty()){
+                    $found=true;
+                    break;
+                }
+            }
+            if(!$found){
+                $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+                return $this->redirect($this->referer());
+            }
+        }
         $articulo = $this->Articulos->get($id, [
             'contain' => ['Modelos', 'Accesorios', 'Asignaciones', 'Devoluciones', 'Lineas']
         ]);
@@ -67,6 +94,10 @@ class ArticulosController extends AppController
      */
     public function add()
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+            return $this->redirect($this->referer());
+        }
         $articulo = $this->Articulos->newEntity();
         if ($this->request->is('post')) {
             $articulo = $this->Articulos->patchEntity($articulo, $this->request->data);
@@ -92,6 +123,10 @@ class ArticulosController extends AppController
      */
     public function edit($id = null)
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+            return $this->redirect($this->referer());
+        }
         $articulo = $this->Articulos->get($id, [
             'contain' => ['Modelos', 'Accesorios']
         ]);
@@ -119,6 +154,10 @@ class ArticulosController extends AppController
      */
     public function delete($id = null)
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            $this->Flash->error(__('Usted no tiene permiso para acceder a la accion solicitada.'));
+            return $this->redirect($this->referer());
+        }
         $this->request->allowMethod(['post', 'delete']);
         $articulo = $this->Articulos->get($id);
         if ($this->Articulos->delete($articulo)) {
