+++ src/Controller/AsignacionesController.php
@@ -2,7 +2,7 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
-
+use Cake\ORM\TableRegistry;
 /**
  * Asignaciones Controller
  *
@@ -18,6 +18,10 @@ class AsignacionesController extends AppController
      */
     public function index()
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+            return $this->redirect($this->referer());
+        }
         $this->paginate = [
             'contain' => ['Procesos', 'Articulos']
         ];
@@ -46,6 +50,22 @@ class AsignacionesController extends AppController
      */
     public function view($id = null)
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            $asig=TableRegistry::get('Asignaciones')->get($id);
+            $pro_tra=TableRegistry::get('ProcesosTrabajadores')->find();
+            if($pro_tra->isEmpty()){
+                $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+                return $this->redirect($this->referer());
+            }
+            $found=false;
+            foreach ($pro_tra as $rowpt)
+                if($rowpt->proceso_id==$asig->proceso_id&&$rowpt->trabajador_id==$this->request->session()->read('Auth.User.trabajador_id'))
+                    $found=true;
+            if(!$found){
+                $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+                return $this->redirect($this->referer());
+            }
+        }
         $asignacion = $this->Asignaciones->get($id, [
             'contain' => ['Procesos', 'Articulos']
         ]);
@@ -61,12 +81,15 @@ class AsignacionesController extends AppController
      */
     public function add()
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+            return $this->redirect($this->referer());
+        }
         $asignacion = $this->Asignaciones->newEntity();
         if ($this->request->is('post')) {
             $asignacion = $this->Asignaciones->patchEntity($asignacion, $this->request->data);
             if ($this->Asignaciones->save($asignacion)) {
                 $this->Flash->success(__('La asignacion ha sido guardada.'));
-
                 return $this->redirect(['action' => 'index']);
             } else {
                 $this->Flash->error(__('La asignacion no pudo ser guardada. Intente nuevamente.'));
@@ -87,6 +110,10 @@ class AsignacionesController extends AppController
      */
     public function edit($id = null)
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+            return $this->redirect($this->referer());
+        }
         $asignacion = $this->Asignaciones->get($id, [
             'contain' => []
         ]);
@@ -115,6 +142,10 @@ class AsignacionesController extends AppController
      */
     public function delete($id = null)
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            $this->Flash->error(__('Usted no tiene permiso para acceder a la accion solicitada.'));
+            return $this->redirect($this->referer());
+        }
         $this->request->allowMethod(['post', 'delete']);
         $asignacion = $this->Asignaciones->get($id);
         if ($this->Asignaciones->delete($asignacion)) {
